Add update checking at start of actions
