+++ agent-package-manager/src/commands/install.ts
@@ -5,9 +5,12 @@ import { InstallOptions } from '../types/index.js';
 import { performToolInstallation } from '../utils/commands/install/toolInstall.js';
 import { getToolInstallDir } from '../utils/shared/path.js';
 import { resolveToolInfo } from '../utils/commands/install/utils.js';
+import checkForUpdates from '../utils/update/index.js';
+
 
 const install = async (toolName: string, options: InstallOptions): Promise<CommandResult> => {
   const spinner = ora(`Installing ${toolName}...`).start();
+  const _checkForUpdates = await checkForUpdates();
   try {
     const { installDir } = await getToolInstallDir();
     if (!installDir) {
